requirements.txt
affine==2.4.0
apispec==6.3.0
apispec-webframeworks==0.5.2
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
click==8.1.7
click-plugins==1.1.1
click-spinner==0.1.10
cligj==0.7.2
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
mercantile==1.2.1
numpy==1.23.5
packaging==23.2
Pillow==10.0.1
pyparsing==3.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
rasterio==1.3.8.post2
shapely==2.0.2
snuggs==1.4.7
SQLAlchemy==2.0.22
toml==0.10.2
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
